rename gammapy/cube/pig_002.py -> gammapy/cube/new.py
