+++ gammapy/cube/__init__.py
@@ -9,4 +9,4 @@ from .utils import *
 from .models import *
 from .cube_pipe import *
 from .basic_cube import *
-from .pig_002 import *
+from .new import *
