+++ gammapy/cube/new.py
